requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart
python-dotenv
requests
pillow
numpy
# PyTorch versión CPU para que Docker no pese 5GB (a menos que uses GPU)
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
gdown